* First part of scrum graphing working, able to generate the day conversion table
DataSeriesFile.hpp: coding convention

TypeIndexModule.hpp: add accessor for the type

ExtentType.cpp: fix compilation warning

perl-client.pl: add TODO for additional testing we need

server.cpp: factor out the code for throwing errors so the server can log the
  error it's sending to the the client.  Tolerate empty extents in the sorted update.
  Refactor so we store extent type pointers everywhere rather than just the name --
  needed the full type to work out what to do when creating a table from an update.
  Handle the case where we do an update and the base table doesn't yet exist.

scrum.pl: remove the sql bits for doing the update, add in the DSS bits.  Copy over
  getTableData from perl-client for help in debugging.

+++ src/base/ExtentType.cpp
@@ -648,7 +648,7 @@ namespace {
 }
 
 const string &ExtentType::fieldTypeToStr(fieldType type) {
-    SINVARIANT(type >= 0 && type < n_field_types);
+    SINVARIANT(type >= 0 && static_cast<uint32_t>(type) < n_field_types);
     return field_types[type];
 }
